refactor(tool-schemas): move todo schemas into conventional files
